- Annoyed by warning noise (mostly shadowing and unused params), fixed the ones that look harmless enough (gone from 1647 to 724). Hopefully didn't break anything.
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@8511 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/sdk/wxscintilla/src/scintilla/include/Platform.h
@@ -69,8 +69,10 @@
 namespace Scintilla {
 #endif
 
-typedef float XYPOSITION;
-typedef double XYACCUMULATOR;
+#include <cstdint>
+
+typedef intptr_t XYPOSITION;
+typedef intptr_t XYACCUMULATOR;
 //#define XYPOSITION int
 
 // Underlying the implementation of the platform classes are platform specific types.
